requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via click
flask==3.1.2
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==25.0
    # via gunicorn
python-dotenv==1.1.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
